requirements.txt
Django<2.0
django-all-access==0.9.0
django-crispy-forms
django-tinymce
django-bleach
django-compressor
django-cleanup
django-debug-toolbar<2.0
django-select2<7.0.0
dj-static
python-dateutil
Pillow
Faker
mypy
django-stubs
django-imagekit
pyocclient
xkcdpass
